+++ Demo/Application/Base/Base View Controllers/BraintreeDemoPaymentButtonBaseViewController.m
@@ -23,8 +23,8 @@
     [self.view addSubview:self.paymentButton];
 
     [NSLayoutConstraint activateConstraints:@[
-        [self.paymentButton.leadingAnchor constraintEqualToAnchor:self.view.leadingAnchor],
-        [self.paymentButton.trailingAnchor constraintEqualToAnchor:self.view.trailingAnchor],
+        [self.paymentButton.leadingAnchor constraintEqualToAnchor:self.view.leadingAnchor constant:20],
+        [self.paymentButton.trailingAnchor constraintEqualToAnchor:self.view.trailingAnchor constant:-20],
         [self.paymentButton.centerYAnchor constraintEqualToAnchor:self.view.centerYAnchor constant:self.centerYConstant],
         [self.paymentButton.heightAnchor constraintEqualToConstant:100.0]
     ]];