+++ Demo/Application/Features/3D Secure/BraintreeDemoThreeDSecurePaymentFlowViewController.h
@@ -1,7 +1,0 @@
-#import <UIKit/UIKit.h>
-
-#import "BraintreeDemoPaymentButtonBaseViewController.h"
-
-@interface BraintreeDemoThreeDSecurePaymentFlowViewController : BraintreeDemoPaymentButtonBaseViewController
-
-@end