+++ Sources/BraintreeCore/Analytics/FPTIBatchData.swift
@@ -30,11 +30,13 @@ struct FPTIBatchData: Codable {
         let errorDescription: String?
         let eventName: String
         let timestamp: String
-        
+        let tenantName: String = "Braintree"
+
         enum CodingKeys: String, CodingKey {
             case errorDescription = "error_desc"
             case eventName = "event_name"
             case timestamp = "t"
+            case tenantName = "tenant_name"
         }
     }
     
@@ -51,7 +53,7 @@ struct FPTIBatchData: Codable {
 
         let clientOS: String = UIDevice.current.systemName + " " + UIDevice.current.systemVersion
 
-        let component = "btmobilesdk"
+        let component = "braintreeclientsdk"
 
         let deviceManufacturer = "Apple"
 
@@ -97,9 +99,7 @@ struct FPTIBatchData: Codable {
         let platform = "iOS"
 
         let sessionID: String
-        
-        let tenantName = "Braintree"
-        
+
         let tokenizationKey: String?
 
         enum CodingKeys: String, CodingKey {
@@ -120,7 +120,6 @@ struct FPTIBatchData: Codable {
             case merchantID = "merchant_id"
             case platform = "platform"
             case sessionID = "session_id"
-            case tenantName = "tenant_name"
             case tokenizationKey = "tokenization_key"
         }
     }