+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Braintree iOS SDK Release Notes
 
+## unreleased
+* BraintreeCore
+  * Fix bug where `type` was always returned as `Unknown` in `fetchPaymentMethodNonces` (fixes #1099)
+
 ## 6.6.0 (2023-08-22)
 * BraintreePayPalNativeCheckout
   * Update PayPalCheckout from 1.0.0 to 1.1.0.