+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## unreleased
 * Send `tenant_name` in `event_params` to PayPal's analytics service (FPTI)
+* Update `component` from `btmobilesdk` to `braintreeclientsdk` for PayPal's analytics service (FPTI)
+* BraintreeCore
+  * Fix bug where `type` was always returned as `Unknown` in `fetchPaymentMethodNonces` (fixes #1099)
 
 ## 6.6.0 (2023-08-22)
 * BraintreePayPalNativeCheckout